--- JetpackGame2D/Assets/Prefabs/laser.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6090ca8d08c0dbb49bcf49a0cf58e726
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
