+++ JetpackGame2D/Assets/Scenes/JetpackGame.unity
@@ -148,7 +148,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 766705727}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0.01, y: 0.38, z: 0.4029404}
+  m_LocalPosition: {x: -9.24, y: 8.49, z: 0.4029404}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -307,134 +307,6 @@ Camera:
   m_OcclusionCulling: 1
   m_StereoConvergence: 10
   m_StereoSeparation: 0.022
---- !u!1 &1442197656
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1442197657}
-  - component: {fileID: 1442197660}
-  - component: {fileID: 1442197659}
-  - component: {fileID: 1442197658}
-  m_Layer: 0
-  m_Name: laser
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &1442197657
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1442197656}
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 10.75, y: 0.84, z: 0}
-  m_LocalScale: {x: 1.929, y: 1.5176, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 3550559020465397415}
-  m_RootOrder: 7
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1442197658
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1442197656}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 917dfc851babd6c4eb04d5a81cdbd5c5, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  laserOnSprite: {fileID: 21300000, guid: fe71038e350b541108b9576ad7add7c6, type: 3}
-  laserOffSprite: {fileID: 21300000, guid: 40d838075b8924cbc9695106def3dbdf, type: 3}
-  toggleInterval: 0.5
-  rotationSpeed: 30
---- !u!61 &1442197659
-BoxCollider2D:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1442197656}
-  m_Enabled: 1
-  m_Density: 1
-  m_Material: {fileID: 0}
-  m_IsTrigger: 1
-  m_UsedByEffector: 0
-  m_UsedByComposite: 0
-  m_Offset: {x: 0, y: 0}
-  m_SpriteTilingProperty:
-    border: {x: 0, y: 0, z: 0, w: 0}
-    pivot: {x: 0.5, y: 0.5}
-    oldSize: {x: 0.39, y: 3.7}
-    newSize: {x: 0.39, y: 3.7}
-    adaptiveTilingThreshold: 0.5
-    drawMode: 0
-    adaptiveTiling: 0
-  m_AutoTiling: 0
-  serializedVersion: 2
-  m_Size: {x: 0.18, y: 3.1}
-  m_EdgeRadius: 0
---- !u!212 &1442197660
-SpriteRenderer:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1442197656}
-  m_Enabled: 1
-  m_CastShadows: 0
-  m_ReceiveShadows: 0
-  m_DynamicOccludee: 1
-  m_StaticShadowCaster: 0
-  m_MotionVectors: 1
-  m_LightProbeUsage: 1
-  m_ReflectionProbeUsage: 1
-  m_RayTracingMode: 0
-  m_RayTraceProcedural: 0
-  m_RenderingLayerMask: 1
-  m_RendererPriority: 0
-  m_Materials:
-  - {fileID: 10754, guid: 0000000000000000f000000000000000, type: 0}
-  m_StaticBatchInfo:
-    firstSubMesh: 0
-    subMeshCount: 0
-  m_StaticBatchRoot: {fileID: 0}
-  m_ProbeAnchor: {fileID: 0}
-  m_LightProbeVolumeOverride: {fileID: 0}
-  m_ScaleInLightmap: 1
-  m_ReceiveGI: 1
-  m_PreserveUVs: 0
-  m_IgnoreNormalsForChartDetection: 0
-  m_ImportantGI: 0
-  m_StitchLightmapSeams: 1
-  m_SelectedEditorRenderState: 0
-  m_MinimumChartSize: 4
-  m_AutoUVMaxDistance: 0.5
-  m_AutoUVMaxAngle: 89
-  m_LightmapParameters: {fileID: 0}
-  m_SortingLayerID: 0
-  m_SortingLayer: 0
-  m_SortingOrder: 0
-  m_Sprite: {fileID: 21300000, guid: fe71038e350b541108b9576ad7add7c6, type: 3}
-  m_Color: {r: 1, g: 1, b: 1, a: 1}
-  m_FlipX: 0
-  m_FlipY: 0
-  m_DrawMode: 0
-  m_Size: {x: 0.39, y: 3.7}
-  m_AdaptiveModeThreshold: 0.5
-  m_SpriteTileMode: 0
-  m_WasSpriteAssigned: 1
-  m_MaskInteraction: 0
-  m_SpriteSortPoint: 0
 --- !u!1 &1523776340
 GameObject:
   m_ObjectHideFlags: 0
@@ -733,8 +605,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 84e6a1da99e8d3e499b63a7a4372faf9, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  forwardMovementSpeed: 7
-  jetpackForce: 50
+  forwardMovementSpeed: 3
+  jetpackForce: 10
 --- !u!50 &1735037761
 Rigidbody2D:
   serializedVersion: 4
@@ -750,7 +622,7 @@ Rigidbody2D:
   m_Mass: 1
   m_LinearDrag: 0
   m_AngularDrag: 0.05
-  m_GravityScale: 2
+  m_GravityScale: 5
   m_Material: {fileID: 0}
   m_Interpolate: 0
   m_SleepingMode: 1
@@ -5817,7 +5689,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1966497069}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0.07, y: 0.34, z: 0.4029404}
+  m_LocalPosition: {x: -9.23, y: 8.4, z: 0.4029404}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -5883,22 +5755,6 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
-    - target: {fileID: 3550559020298073346, guid: 5d3d1f50b64f2414bbfe80cb2ff99a26, type: 3}
-      propertyPath: m_LocalScale.x
-      value: 1.426629
-      objectReference: {fileID: 0}
-    - target: {fileID: 3550559020298073346, guid: 5d3d1f50b64f2414bbfe80cb2ff99a26, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: -6.2957
-      objectReference: {fileID: 0}
-    - target: {fileID: 3550559020613399663, guid: 5d3d1f50b64f2414bbfe80cb2ff99a26, type: 3}
-      propertyPath: m_LocalScale.x
-      value: 22.810034
-      objectReference: {fileID: 0}
-    - target: {fileID: 3550559020613399663, guid: 5d3d1f50b64f2414bbfe80cb2ff99a26, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 1.7084
-      objectReference: {fileID: 0}
     - target: {fileID: 3550559020888591194, guid: 5d3d1f50b64f2414bbfe80cb2ff99a26, type: 3}
       propertyPath: m_RootOrder
       value: 0
@@ -5947,34 +5803,5 @@ PrefabInstance:
       propertyPath: m_Name
       value: House
       objectReference: {fileID: 0}
-    - target: {fileID: 3550559021840558254, guid: 5d3d1f50b64f2414bbfe80cb2ff99a26, type: 3}
-      propertyPath: m_LocalScale.x
-      value: 22.847715
-      objectReference: {fileID: 0}
-    - target: {fileID: 3550559021840558254, guid: 5d3d1f50b64f2414bbfe80cb2ff99a26, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 1.7063
-      objectReference: {fileID: 0}
-    - target: {fileID: 3550559022065974387, guid: 5d3d1f50b64f2414bbfe80cb2ff99a26, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 1.0083
-      objectReference: {fileID: 0}
-    - target: {fileID: 3550559022065974387, guid: 5d3d1f50b64f2414bbfe80cb2ff99a26, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0.15
-      objectReference: {fileID: 0}
-    - target: {fileID: 3550559022110669552, guid: 5d3d1f50b64f2414bbfe80cb2ff99a26, type: 3}
-      propertyPath: m_LocalScale.x
-      value: 1.7197756
-      objectReference: {fileID: 0}
-    - target: {fileID: 3550559022110669552, guid: 5d3d1f50b64f2414bbfe80cb2ff99a26, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 9.024
-      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 5d3d1f50b64f2414bbfe80cb2ff99a26, type: 3}
---- !u!4 &3550559020465397415 stripped
-Transform:
-  m_CorrespondingSourceObject: {fileID: 3550559020888591194, guid: 5d3d1f50b64f2414bbfe80cb2ff99a26, type: 3}
-  m_PrefabInstance: {fileID: 3550559020465397414}
-  m_PrefabAsset: {fileID: 0}
