--- JetpackGame2D/Assets/Scripts/Character.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d8e50da19f300a042997194aaf0364cc
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
