--- JetpackGame2D/Assets/Scripts/LaserScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 917dfc851babd6c4eb04d5a81cdbd5c5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
